backend/requirements.txt
antlr4-python3-runtime
google-genai
python-dotenv
fastapi
uvicorn
pydantic
python-multipart